Update index.md
I talked to Johan. He suggested to add some more detail on the landing page, to explain how to navigate the site a bit better. 
Here we should also explain what a .json file is and how to open .md files (another problem that Johan had, we rely on them in the dowload; worst case we need to add a txt version?)
 what is a good cross-platform way to do this? requiering useres to downloading atom is probably a bit of an overkill.