Fix the plant list
In the last release of the dataset we introduced characters to the column headers that are incompatible with the framework of the webpage (and likely with other tools as well): "[", "]", ";", "/".

As a hot fix, we copy the dataset here and replace those incompatible characters. This should be removed once we release a new dataset without the incompatible characters.